+++ api_app/models/schemas/airlock_request.py
@@ -1,8 +1,9 @@
+import uuid
 from datetime import datetime
 from typing import List
 from pydantic import BaseModel, Field
 from models.domain.airlock_resource import AirlockResourceType
-from models.domain.airlock_request import AirlockRequest, AirlockRequestType
+from models.domain.airlock_request import AirlockActions, AirlockRequest, AirlockRequestType
 
 
 def get_sample_airlock_request(workspace_id: str, airlock_request_id: str) -> dict:
@@ -18,6 +19,13 @@ def get_sample_airlock_request(workspace_id: str, airlock_request_id: str) -> di
     }
 
 
+def get_sample_airlock_request_with_allowed_user_actions(workspace_id: str) -> dict:
+    return {
+        "airlockRequest": get_sample_airlock_request(workspace_id, str(uuid.uuid4())),
+        "allowed_user_actions": [AirlockActions.Cancel, AirlockActions.Review, AirlockActions.Submit],
+    }
+
+
 class AirlockRequestInResponse(BaseModel):
     airlockRequest: AirlockRequest
 
@@ -29,15 +37,25 @@ class AirlockRequestInResponse(BaseModel):
         }
 
 
-class AirlockRequestInList(BaseModel):
-    airlockRequests: List[AirlockRequest] = Field([], title="Airlock Requests")
+class AirlockRequestWithAllowedUserActions(BaseModel):
+    airlockRequest: AirlockRequest = Field([], title="Airlock Request")
+    allowed_user_actions: List[str] = Field([], title="actions that the requesting user can do on the request")
+
+    class Config:
+        schema_extra = {
+            "example": get_sample_airlock_request_with_allowed_user_actions("933ad738-7265-4b5f-9eae-a1a62928772e"),
+        }
+
+
+class AirlockRequestWithAllowedUserActionsInList(BaseModel):
+    airlockRequests: List[AirlockRequestWithAllowedUserActions] = Field([], title="Airlock Requests")
 
     class Config:
         schema_extra = {
             "example": {
                 "airlock_requests": [
-                    get_sample_airlock_request("933ad738-7265-4b5f-9eae-a1a62928772e", "121e921f-a4aa-44b3-90a9-e8da030495ef"),
-                    get_sample_airlock_request("123ad738-1234-4b5f-9eae-a1a62928772e", "457e921f-a4aa-44b3-90a9-e8da030412ac"),
+                    get_sample_airlock_request_with_allowed_user_actions("933ad738-7265-4b5f-9eae-a1a62928772e"),
+                    get_sample_airlock_request_with_allowed_user_actions("933ad738-7265-4b5f-9eae-a1a62928772e")
                 ]
             }
         }